rename route-specific components
