+++ src/modules/routing/components/routeDetails/index.js
@@ -1,15 +1,15 @@
 import { RouteDetails } from './RouteDetails';
-import { RoutingChart } from './RoutingChart';
-import { RoutingWarnings } from './RoutingWarnings';
+import { RouteChart } from './RouteChart';
+import { RouteWarnings } from './RouteWarnings';
 
 if (!window.customElements.get(RouteDetails.tag)) {
 	window.customElements.define(RouteDetails.tag, RouteDetails);
 }
 
-if (!window.customElements.get(RoutingChart.tag)) {
-	window.customElements.define(RoutingChart.tag, RoutingChart);
+if (!window.customElements.get(RouteChart.tag)) {
+	window.customElements.define(RouteChart.tag, RouteChart);
 }
 
-if (!window.customElements.get(RoutingWarnings.tag)) {
-	window.customElements.define(RoutingWarnings.tag, RoutingWarnings);
+if (!window.customElements.get(RouteWarnings.tag)) {
+	window.customElements.define(RouteWarnings.tag, RouteWarnings);
 }
