+++ src/modules/routing/index.js
@@ -2,7 +2,7 @@ import './i18n';
 
 import { CategoryBar } from './components/categoryBar/CategoryBar';
 import { FeedbackBanner } from './components/feedbackBanner/FeedbackBanner';
-import { RoutingInfo } from './components/routingInfo/RoutingInfo';
+import { RouteInfo } from './components/routeInfo/RouteInfo';
 import { Waypoints } from './components/waypoints/Waypoints';
 import { WaypointItem } from './components/waypoints/WaypointItem';
 import './components/routeDetails';
@@ -13,8 +13,8 @@ if (!window.customElements.get(CategoryBar.tag)) {
 if (!window.customElements.get(FeedbackBanner.tag)) {
 	window.customElements.define(FeedbackBanner.tag, FeedbackBanner);
 }
-if (!window.customElements.get(RoutingInfo.tag)) {
-	window.customElements.define(RoutingInfo.tag, RoutingInfo);
+if (!window.customElements.get(RouteInfo.tag)) {
+	window.customElements.define(RouteInfo.tag, RouteInfo);
 }
 if (!window.customElements.get(Waypoints.tag)) {
 	window.customElements.define(Waypoints.tag, Waypoints);
